# cms-backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Full-Stack CMS API",
    "description": "Enterprise-grade REST API for a full-stack CMS with JWT auth and RBAC",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/blogs": {
      "post": {
        "tags": ["blogs"],
        "summary": "Create Blog Post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slug": { "type": "string" },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published"]
                  },
                  "translations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "languageCode": { "type": "string" },
                        "title": { "type": "string" },
                        "content": { "type": "string" }
                      }
                    }
                  }
                }
              },
              "examples": {
                "EnglishAndArabic": {
                  "value": {
                    "slug": "getting-started-with-nextjs",
                    "status": "published",
                    "translations": [
                      {
                        "languageCode": "en",
                        "title": "Getting Started with Next.js",
                        "content": "Next.js is a React framework for building fast, modern web applications. In this comprehensive guide, we'll explore the foundational concepts, best practices, and advanced techniques to help you build scalable applications. Learn about server-side rendering, static generation, API routes, and more."
                      },
                      {
                        "languageCode": "ar",
                        "title": "البدء مع Next.js",
                        "content": "Next.js هو إطار عمل React لبناء تطبيقات ويب سريعة وحديثة. في هذا الدليل الشامل، سوف نستكشف المفاهيم الأساسية وأفضل الممارسات والتقنيات المتقدمة لمساعدتك في بناء تطبيقات قابلة للتوسع. تعرف على العرض من جانب الخادم والتوليد الثابت ومسارات واجهة برمجة التطبيقات والمزيد."
                      }
                    ]
                  }
                },
                "TechTutorial": {
                  "value": {
                    "slug": "mastering-typescript-generics",
                    "status": "draft",
                    "translations": [
                      {
                        "languageCode": "en",
                        "title": "Mastering TypeScript Generics",
                        "content": "TypeScript generics allow you to create flexible and reusable code components. This tutorial covers generic functions, interfaces, classes, and constraints. Understand how to write type-safe code that works with multiple data types while maintaining full type checking."
                      },
                      {
                        "languageCode": "ar",
                        "title": "إتقان النوع العام في TypeScript",
                        "content": "تسمح النماذج العامة في TypeScript بإنشاء مكونات كود مرنة وقابلة لإعادة الاستخدام. يغطي هذا البرنامج التعليمي الوظائف والواجهات والفئات العامة والقيود. افهم كيفية كتابة كود آمن من النوع يعمل مع أنواع بيانات متعددة مع الحفاظ على التحقق الكامل من النوع."
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog created"
          }
        }
      },
      "get": {
        "tags": ["blogs"],
        "summary": "List Blog Posts (Admin)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 10, "maximum": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Blogs retrieved"
          }
        }
      }
    },
    "/blogs/public": {
      "get": {
        "tags": ["blogs"],
        "summary": "List Published Blogs (Public)",
        "security": [],
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "schema": { "type": "string", "default": "en" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 10, "maximum": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Published blogs retrieved"
          }
        }
      }
    },
    "/blogs/{id}": {
      "get": {
        "tags": ["blogs"],
        "summary": "Get Blog by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog retrieved"
          }
        }
      },
      "patch": {
        "tags": ["blogs"],
        "summary": "Update Blog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slug": { "type": "string" },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published"]
                  },
                  "translations": { "type": "array" }
                }
              },
              "examples": {
                "UpdateToPublished": {
                  "value": {
                    "status": "published",
                    "slug": "getting-started-with-nextjs",
                    "translations": [
                      {
                        "languageCode": "en",
                        "title": "Getting Started with Next.js - Updated",
                        "content": "Next.js is a React framework for building fast, modern web applications. This updated version includes new sections on App Router, Server Actions, and performance optimization techniques."
                      },
                      {
                        "languageCode": "ar",
                        "title": "البدء مع Next.js - تم التحديث",
                        "content": "Next.js هو إطار عمل React لبناء تطبيقات ويب سريعة وحديثة. تتضمن هذه الإصدارة المحدثة أقسام جديدة حول جهاز التوجيه والإجراءات من جانب الخادم وتقنيات تحسين الأداء."
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog updated"
          }
        }
      },
      "delete": {
        "tags": ["blogs"],
        "summary": "Delete Blog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog deleted"
          }
        }
      }
    },
...
  }
}